requirement.txt
asgiref==3.7.2
Django==5.0.1
django-shortcuts==1.6
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
